backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
firebase_admin==6.8.0